Rename tests/iam to tests/services to match new naming.
