+++ tests/services/api_tests/model_test.py
@@ -22,8 +22,8 @@ from google.cloud.forseti.services.inventory.service import GrpcInventoryFactory
 from google.cloud.forseti.services.model.service import GrpcModellerFactory
 from google.cloud.forseti.services.dao import ModelManager
 
-from tests.iam.api_tests.api_tester import ModelTestRunner
-from tests.iam.utils.db import create_test_engine
+from tests.services.api_tests.api_tester import ModelTestRunner
+from tests.services.utils.db import create_test_engine
 from tests.unittest_utils import ForsetiTestCase
