+++ tests/services/api_tests/playground_test.py
@@ -22,9 +22,9 @@ from google.cloud.forseti.services.inventory.service import GrpcInventoryFactory
 from google.cloud.forseti.services.model.service import GrpcModellerFactory
 from google.cloud.forseti.services.dao import ModelManager
 
-from tests.iam.api_tests.api_tester import ApiTestRunner
-from tests.iam.utils.db import create_test_engine, cleanup
-from tests.iam.utils.mock import MockServerConfig
+from tests.services.api_tests.api_tester import ApiTestRunner
+from tests.services.utils.db import create_test_engine, cleanup
+from tests.services.utils.mock import MockServerConfig
 from tests.unittest_utils import ForsetiTestCase
