+++ tests/services/cli_test.py
