+++ tests/services/inventory/crawling_test.py
@@ -16,13 +16,12 @@
 
 import unittest
 
-
-from tests.unittest_utils import ForsetiTestCase
 from google.cloud.forseti.services.inventory.base.storage import Memory as MemoryStorage
 from google.cloud.forseti.services.inventory.base.progress import Progresser
 from google.cloud.forseti.services.inventory.crawler import run_crawler
 from google.cloud.forseti.services.server import InventoryConfig
-from tests.iam.utils.gcp_env import gcp_configured, gcp_env
+from tests.services.utils.gcp_env import gcp_configured, gcp_env
+from tests.unittest_utils import ForsetiTestCase
 
 
 class NullProgresser(Progresser):
