+++ tests/services/inventory/storage_test.py
@@ -15,13 +15,13 @@
 """ Unit Tests: Inventory storage for IAM Explain. """
 
 import unittest
-from tests.unittest_utils import ForsetiTestCase
 from sqlalchemy import event
 
 from google.cloud.forseti.services.inventory.storage import Storage, initialize
 from google.cloud.forseti.services import db
 from google.cloud.forseti.services.inventory.base.resources import Resource
-from tests.iam.utils.db import create_test_engine
+from tests.services.utils.db import create_test_engine
+from tests.unittest_utils import ForsetiTestCase
 
 
 class ResourceMock(Resource):
