+++ tests/services/model/importer/importer_test.py
@@ -19,12 +19,12 @@ from shutil import copyfile
 import tempfile
 import unittest
 
-from tests.unittest_utils import ForsetiTestCase
 from google.cloud.forseti.services.dao import ModelManager, create_engine
 from google.cloud.forseti.services.inventory.storage import InventoryState
 from google.cloud.forseti.services.model.importer import importer
-from tests.iam.utils.gcp_env import gcp_configured, gcp_env
-from tests.iam.utils.protect import copy_file_decrypt
+from tests.services.utils.gcp_env import gcp_configured, gcp_env
+from tests.services.utils.protect import copy_file_decrypt
+from tests.unittest_utils import ForsetiTestCase
 
 
 class ServiceConfig(object):
