+++ tests/services/model_manager_test.py
@@ -17,11 +17,11 @@
 import os
 import unittest
 
-from tests.iam.utils.db import create_test_engine_with_file
-from tests.unittest_utils import ForsetiTestCase
 
 from google.cloud.forseti.services.dao import ModelManager
 from google.cloud.forseti.common.util.threadpool import ThreadPool
+from tests.services.utils.db import create_test_engine_with_file
+from tests.unittest_utils import ForsetiTestCase
 
 
 class ModelManagerTest(ForsetiTestCase):
